Restructured handlers folder structure
