+++ api/routes/routes.go
@@ -2,7 +2,7 @@ package routes
 
 import (
 	"database/sql"
-	"web-app/api/routes/handlers"
+	"web-app/api/handlers/user_handler"
 
 	"github.com/labstack/echo/v4"
 )
